+++ hierarchical.h
@@ -77,12 +77,12 @@ namespace hicma {
 
     void trsm(const Node& A, const char& uplo) override;
 
-    void gemm(const Node& A, const Node& B, const int& alpha=-1, const int& beta=1) override;
+    void gemm(const Node& A, const Node& B, const double& alpha=-1, const double& beta=1) override;
 
     void gemm_row(
                   const Hierarchical& A, const Hierarchical& B,
                   const int& i, const int& j, const int& k_min, const int& k_max,
-                  const int& alpha, const int& beta);
+                  const double& alpha, const double& beta);
   };
 }
 #endif
